Добавляет функцию перезагрузки карты
