+++ js/constant.js
@@ -31,11 +31,13 @@ const START_POINTS_OBJECT = {
   lng: TOKYO_LONGITUDE,
 };
 
+const MODAL_SHOW_TIME = 5000;
+
 const ERROR_MESSAGE = 'Не удалось отправить форму. Попробуйте ещё раз';
 
 export {
   MIN_ELEMENT, MIN_POSITIVE_NUMBER, COUNT_OF_PINS, MIN_LOCATION_X, MIN_LOCATION_Y,
   MAX_LOCATION_X, MAX_LOCATION_Y, MAX_PRICE, MAX_COUNT_OF_AVATARS, MAX_COUNT_OF_DECIMAL_NUMBERS,
   MAX_FEATURES, MAX_PHOTOS, MIN_PRICE, TOKYO_LATITUDE, TOKYO_LONGITUDE, MAIN_PIN, PIN, START_POINTS,
-  START_POINTS_OBJECT, ERROR_MESSAGE
+  START_POINTS_OBJECT, ERROR_MESSAGE, MODAL_SHOW_TIME
 };
